refactor out HIGChildValidator
