+++ packages/react-hig/src/elements/components/GlobalNav/GlobalNav.js
@@ -16,9 +16,9 @@ limitations under the License.
 */
 import * as HIG from 'hig.web';
 import * as PropTypes from 'prop-types';
-import React from 'react';
 
 import HIGElement from '../../HIGElement';
+import HIGChildValidator from '../../HIGChildValidator';
 import createComponent from '../../../adapters/createComponent';
 
 import ContainerComponent, { Container } from './Container';
@@ -128,32 +128,9 @@ class GlobalNav extends HIGElement {
 
 const GlobalNavComponent = createComponent(GlobalNav);
 
-const validChildren = [SideNavComponent, ContainerComponent];
-
 GlobalNavComponent.propTypes = {
   sideNavOpen: PropTypes.bool,
-  children: (props, propName, componentName) => {
-    const prop = props[propName];
-
-    let error = null;
-    React.Children.forEach(prop, function(child) {
-      const getDisplayName = Component =>
-        Component.displayName ||
-        Component.name ||
-        (typeof Component === 'string' ? Component : 'Component');
-
-      const childDisplayName = getDisplayName(child.type || child);
-
-      if (validChildren.indexOf(child.type) === -1) {
-        error = new Error(
-          `'${childDisplayName}' is not a valid child of ${componentName}. Children should be of type '${validChildren
-            .map(c => c.displayName)
-            .join(', ')}'.`
-        );
-      }
-    });
-    return error;
-  }
+  children: HIGChildValidator([SideNavComponent, ContainerComponent])
 };
 
 GlobalNavComponent.Container = ContainerComponent;
